requirements.txt
amqp==5.1.1
anyio==3.7.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
billiard==4.1.0
boto3==1.28.26
botocore==1.31.26
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.5
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
cryptography==41.0.2
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.4
django-advance-thumbnail==1.0.0
django-allauth==0.54.0
django-celery-beat==2.5.0
django-cors-headers==4.2.0
django-debug-toolbar==4.1.0
django-filter==23.2
django-simple-history==3.3.0
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
einops==0.6.1
encodec==0.1.1
exceptiongroup==1.1.2
executing==1.2.0
filelock==3.12.2
fsspec==2023.6.0
funcy==2.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
idna==3.4
inflection==0.5.1
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
numpy==1.25.2
oauthlib==3.2.2
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.7.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
redis==4.6.0
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
safetensors==0.3.2
scipy==1.11.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.2
suno-bark @ git+https://github.com/suno-ai/bark.git@56b0ba13f7c281cbffa07ea9abf7b30273a60b6a
sympy==1.12
tokenizers==0.13.3
# torch==2.0.1
torchaudio==2.0.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.31.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.1
uvloop==0.17.0
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
